Partially Completed Part2 of Exercise 14

+++ posts/views.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, flash, url_for, redirect, request
+from flask import Blueprint, render_template, flash, url_for, redirect, request, abort
 from flask_login import login_required, current_user
 from config import db, Post
 from posts.forms import PostForm
@@ -9,12 +9,18 @@ posts_bp = Blueprint('posts', __name__, template_folder='templates')
 @posts_bp.route('/posts')
 @login_required
 def posts():
+    if current_user.role in ['db_admin', 'sec_admin']:
+        flash("You are not authorized to view posts.", "danger")
+        return redirect(url_for('accounts.forbidden_error_page', error='Access Denied'))
     all_posts = Post.query.order_by(desc('id')).all()
     return render_template('posts/posts.html', posts=all_posts)
 
 @posts_bp.route('/create', methods=('GET', 'POST'))
 @login_required
 def create():
+    if current_user.role in ['db_admin', 'sec_admin']:
+        flash("You are not authorized to create posts.", "danger")
+        return redirect(url_for('accounts.forbidden_error_page', error='Access Denied'))
     form = PostForm()
     if form.validate_on_submit():
         new_post = Post(title=form.title.data, body=form.body.data, userid=current_user.id)
@@ -28,18 +34,13 @@ def create():
 @login_required
 def update(id):
     post_to_update = Post.query.filter_by(id=id).first()
-    if not post_to_update:
-        flash("Post does not exist.", "danger")
-        return redirect(url_for('posts.posts'))
-
-    # Restrict update to post owner
-    if post_to_update.userid != current_user.id:
-        flash("You are not authorized to update this post.", "danger")
-        return redirect(url_for('posts.posts'))
+    if not post_to_update or post_to_update.userid != current_user.id:
+        abort(403)
 
     form = PostForm()
     if form.validate_on_submit():
         post_to_update.update(title=form.title.data, body=form.body.data)
+        db.session.commit()
         flash('Post updated', category='success')
         return redirect(url_for('posts.posts'))
 
@@ -51,14 +52,10 @@ def update(id):
 @login_required
 def delete(id):
     post_to_delete = Post.query.filter_by(id=id).first()
-    if not post_to_delete:
-        flash("Post does not exist.", "danger")
-        return redirect(url_for('posts.posts'))
 
-    # Restrict delete to post owner
-    if post_to_delete.userid != current_user.id:
+    if not post_to_delete or post_to_delete.userid != current_user.id:
         flash("You are not authorized to delete this post.", "danger")
-        return redirect(url_for('posts.posts'))
+        return redirect(url_for('accounts.forbidden_error_page', error='Access Denied'))
 
     db.session.delete(post_to_delete)
     db.session.commit()
